backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pdfminer.six
httpx
faiss-cpu
numpy
pymupdf
Pillow
slowapi
paddleocr
